feat: secondary button, refactor Cta button
